UKGT: Start using a common old measure data structure.
